requirements.txt
numpy~=1.19.2
pandas~=1.1.5
pytorch~=1.10.2
pytorch-cuda~=11.8
dgl~=1.0.1.cu116
scikit-learn~=0.24.2
scipy~=1.5.2
xlsxwriter~=3.2.0
matplotlib~=3.3.4